+++ packages/app-rfi/src/components/steps/questions/GdprConsent.js
@@ -14,27 +14,27 @@ export const GdprConsent = ({ gaData }) => {
     formik: { values },
   } = useRfiContext();
 
-  const gdprWording1 = (
+  const gdprWordingGround = (
     <>
       By submitting my information, I consent to ASU contacting me about
-      education services using email, direct mail, SMS/texting and digital
+      educational services using email, direct mail, SMS/texting and digital
       platforms. Message and data rates may apply. Consent is not required to
-      receive services, and I can withdraw consent by contacting ASU at{" "}
+      receive services, and I can unsubscribe at any time by contacting{" "}
       <a href="mailto:UnsubFutureStudentComm@asu.edu">
         UnsubFutureStudentComm@asu.edu
-      </a>{" "}
-      or as described in communications I receive. I consent to ASU’s{" "}
-      <a href="https://asuonline.asu.edu/text-terms/">
+      </a>
+      . I consent to ASU’s{" "}
+      <a href="https://asu.edu/about/text-message-terms">
         mobile terms and conditions
-      </a>{" "}
-      and{" "}
-      <a href="https://asuonline.asu.edu/web-analytics-privacy-2/">
+      </a>
+      , and{" "}
+      <a href="https://asu.edu/about/web-analytics-privacy">
         Privacy Statements
       </a>
       , including the European Supplement.
     </>
   );
-  const gdprWording2 = (
+  const gdprWordingOnline = (
     <>
       By submitting my information, I consent to ASU contacting me about
       educational services using automated calls, prerecorded voice messages,
@@ -56,7 +56,7 @@ export const GdprConsent = ({ gaData }) => {
   return (
     <div className="rfi-consent">
       <div className="rfi-consent-wording">
-        {values.Campus !== KEY.ONLINE ? gdprWording1 : gdprWording2}
+        {values.Campus === KEY.ONLINE ? gdprWordingOnline : gdprWordingGround}
       </div>
       <RfiCheckboxSingle
         id="GdprConsent"