+++ packages/component-carousel/__mocks__/data/props-mock.js
@@ -1,17 +1,7 @@
+import { imageArray } from "../../../../shared/assets";
 import { testimonialWithImage } from "../../src/components/TestimonialCarousel/examples";
 
-const imageFormats = [
-  "500x400",
-  "300x400",
-  "400x400",
-  "200x200",
-  "300x400",
-  "300x400",
-  "400x400",
-  "300x400",
-];
-
-const getImageFormat = index => imageFormats[index];
+const getImageFormat = index => imageArray[index];
 
 const cardCarouselItems = [];
 for (let index = 0; index < 8; index += 1) {
@@ -23,17 +13,16 @@ for (let index = 0; index < 8; index += 1) {
     amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt
     ut labore et dolore magna aliqua eiusmod tempo.`;
 
-  const imageSource =
+  const imageSource = index === 5 ? "https://br0ken" : getImageFormat(index);
+  const imageAltText =
     index === 5
-      ? "https://br0ken"
-      : `https://source.unsplash.com/random/${getImageFormat(index)}?a=${
-          index - 1
-        }`;
+      ? "Intentionally broken image. Broken images look like this!"
+      : "Card image cap";
 
   cardCarouselItems.push({
     id: index,
     imageSource,
-    imageAltText: "Card image cap",
+    imageAltText,
     title: `Card ${index + 1}`,
     content,
     buttons: [
@@ -56,31 +45,31 @@ for (let index = 0; index < 8; index += 1) {
 const imageCarouselItems = [
   {
     id: 1,
-    imageSource: "https://source.unsplash.com/random/800x400?a=1",
+    imageSource: imageArray[1],
     imageAltText:
       "Random image with caption below. REPLACE with appropriate alt text for accessibility.",
   },
   {
     id: 2,
-    imageSource: "https://source.unsplash.com/random/400x200?a=2",
+    imageSource: imageArray[2],
     imageAltText:
       "Random image with caption below. REPLACE with appropriate alt text for accessibility.",
   },
   {
     id: 3,
-    imageSource: "https://source.unsplash.com/random/600x400?a=3",
+    imageSource: imageArray[3],
     imageAltText:
       "Random image with caption below. REPLACE with appropriate alt text for accessibility.",
   },
   {
     id: 4,
-    imageSource: "https://source.unsplash.com/random/600x600?a=4",
+    imageSource: imageArray[4],
     imageAltText:
       "Random image with caption below. REPLACE with appropriate alt text for accessibility.",
   },
   {
     id: 5,
-    imageSource: "https://source.unsplash.com/random/600x500?a=5",
+    imageSource: imageArray[5],
     imageAltText:
       "Random image with caption below. REPLACE with appropriate alt text for accessibility.",
   },
@@ -90,9 +79,7 @@ const imageGalleryCarouselItems = [];
 for (let index = 0; index < 8; index += 1) {
   imageGalleryCarouselItems.push({
     id: index,
-    imageSource: `https://source.unsplash.com/random/${getImageFormat(
-      index
-    )}?a=${index - 1}`,
+    imageSource: imageArray[index],
     imageAltText:
       "Random image with caption below. REPLACE with appropriate alt text for accessibility.",
   });