+++ packages/components-core/.storybook/preview.js
@@ -1,4 +1,5 @@
 import "@asu/unity-bootstrap-theme/src/scss/unity-bootstrap-theme.bundle.scss";
+import "../../../node_modules/bootstrap/dist/js/bootstrap.bundle.min.js";
 
 const parameters = {
   actions: { argTypesRegex: "^on[A-Z].*" },