+++ packages/unity-bootstrap-theme/stories/atoms/anchor-menu/anchor-menu.examples.stories.js
@@ -1,5 +1,6 @@
 import React from "react";
 
+import { imageName } from "../../../../../shared/assets";
 import { htmlRootDecorator } from "../../../../../shared/components/Layout";
 import { initializeAnchorMenu as initFunc } from "./anchor-menu";
 
@@ -27,7 +28,7 @@ return (
         <div className="uds-hero-lg has-btn-row">
           <img
             className="hero"
-            src="https://source.unsplash.com/WLUHO9A_xik/1920x1000"
+            src={ imageName.hero02 }
             alt="Sample placeholder image."
             width="1920"
             height="1000"
@@ -56,7 +57,7 @@ return (
         >
           <div className="container">
             <div className="uds-anchor-menu-wrapper">
-              <h4
+              <h2
                 data-bs-toggle="collapse"
                 data-bs-target="#collapseExample"
                 aria-expanded="false"
@@ -67,7 +68,7 @@ return (
                 data-ga="On this page"
               >
                 On This Page: <span className="fas fa-chevron-down"></span>
-              </h4>
+              </h2>
               <div id="collapseExample" className="card card-body collapse">
                 <nav className="nav" aria-label="Same Page">
                   <a